Economics Data

This dataset contains economic data, including personal consumption expenditures, population, personal savings rate, median duration of unemployment, and the number of unemployed individuals, recorded monthly from July 1967.

Usage

economics

Format

A tibble with 574 rows and 6 variables:

date

Date. The date of the observation.

pce

Numeric. Personal consumption expenditures, in billions of dollars.

pop

Numeric. Total population, in thousands.

psavert

Numeric. Personal savings rate, as a percentage.

uempmed

Numeric. Median duration of unemployment, in weeks.

unemploy

Numeric. Number of unemployed individuals, in thousands.

Source

This dataset is sourced from the Federal Reserve Economic Data (FRED) database. https://fred.stlouisfed.org/

Examples

data(economics)
head(economics)
#> # A tibble: 6 × 6
#>   date         pce    pop psavert uempmed unemploy
#>   <date>     <dbl>  <dbl>   <dbl>   <dbl>    <dbl>
#> 1 1967-07-01  507. 198712    12.6     4.5     2944
#> 2 1967-08-01  510. 198911    12.6     4.7     2945
#> 3 1967-09-01  516. 199113    11.9     4.6     2958
#> 4 1967-10-01  512. 199311    12.9     4.9     3143
#> 5 1967-11-01  517. 199498    12.8     4.7     3066
#> 6 1967-12-01  525. 199657    11.8     4.8     3018