Example Dataset: Fruit Data

This dataset contains information about different types of fruits in various baskets over multiple years. It includes details such as the fruit type, the value, and the color of the fruit.

Usage

fruit

Format

A tibble with 16 rows and 6 variables:

id

Integer. A unique identifier for each record.

year

Character. The year the data was recorded.

basket

Integer. The basket number containing the fruits.

fruit

Character. The type of fruit.

value

Integer. The value associated with the fruit.

color

Character. The color of the fruit.

Examples

data(fruit)
head(fruit)
#> # A tibble: 6 × 6
#>      id year  basket fruit    value color 
#>   <int> <chr>  <int> <chr>    <int> <chr> 
#> 1     1 2019       1 apples    3840 red   
#> 2     2 2019       1 bananas   1920 green 
#> 3     3 2019       2 cherries   960 blue  
#> 4     4 2019       2 dates      400 orange
#> 5     5 2018       1 apples    1600 red   
#> 6     6 2018       1 bananas   1440 green