Add Arrows to a SveltePlot Chart

Adds arrows to a SveltePlot chart to illustrate directions, trends, or to point out specific data points. This function enhances the chart's ability to convey insights to the viewer.

Usage

sp_add_arrows(
  sp,
  x_start,
  x_end,
  y_start,
  y_end,
  arrow_head_type = NULL,
  size = NULL,
  color = "black",
  curvature = 1e-05,
  direction = "upward",
  arrow_head = NULL
)

Arguments

sp

A SveltePlot htmlwidget object to which arrows will be added.

x_start

Numeric vector specifying the starting x-coordinates of the arrows.

x_end

Numeric vector specifying the ending x-coordinates of the arrows.

y_start

Numeric vector specifying the starting y-coordinates of the arrows.

y_end

Numeric vector specifying the ending y-coordinates of the arrows.

arrow_head_type

Character vector specifying the type of arrow head. Can be customized to suit different visualization needs.

size

Numeric vector specifying the size of the arrows.

color

Character vector specifying the color of the arrows. Default is "black".

curvature

Numeric vector specifying the curvature of the arrows. This is useful for creating curved arrows that can more naturally point between two points on the chart.

direction

Character vector specifying the direction of the arrow. Valid options are "upward" or "downward". Default is "upward".

arrow_head

Also not sure what it does.

Examples

library(SveltePlots)
data("purchases")

sp(
  data = purchases_age, type = "line",
  mapping = spaes(x = adjusted_timestamp, y = revenue_roll, group = age),
  colors = c("red", "green", "blue"),
  combine_same_groups = TRUE
) |>
  sp_add_series(
    data = purchases_age,
    mapping = spaes(x = adjusted_timestamp, y = revenue, group = age),
    type = "points",
    alpha = 0.4,
    tooltip = FALSE,
    include_legend = FALSE
  ) |>
    sp_add_series(
    purchases_age[purchases_age$revenue == max(purchases_age$revenue), ],
    mapping = spaes(x = adjusted_timestamp, y = revenue, group = age),
    type = "points",
    size = 5,
    tooltip = FALSE
    ) |>
  sp_add_segments(
    x_start = "2023-08-31",
    x_end = "2023-09-05",
    y_start = "auto",
     y_end ="auto",
    type = "rect",
    opacity = 0.2,
    background_color = "black",
    text_color = "white",
    show_legend = TRUE,
    legend_text = "Labor Day Bundle",
    tooltip = "Revenue: <strong>235</strong>"
  ) |>
  sp_add_arrows(
    x_start = c("2023-10-01", "2023-08-01"),
    x_end = c("2023-09-05", "2023-08-31"),
    y_start = c(1200, 1200),
    y_end = c(1000, 1370),
    arrow_head = c(0, 0),
    size = c(200, 200),
    curvature = c(0.2, 0.5),
    direction = c("downward", "downward"),
    color = c("black", "black"),
    arrow_head_type = c("triangle", "triangle")
  ) |>
  sp_add_text(
    x = c("2023-07-10", "2023-10-01"),
    y = c(1300, 1220),
    text = c(
      "This was the highest revenue day \n due to the labor day bundle",
      "Labor Day Bundle Run Window"
    )
  )