Walmart Weekly Sales Data

Weekly Sales Data

Usage

walmart_sales_weekly

Format

A tibble with 1,001 rows and 17 variables:

id

Unique identifier for the combination of Store and Dept, a factor

Store

Store number, a numeric value

Dept

Department number, a numeric value

Date

Date of the observation, in Date format

Weekly_Sales

Sales for the given department in the given store, a numeric value

IsHoliday

Indicator of whether the week is a special holiday week, a logical value

Type

Type of store, a character string

Size

Size of the store, a numeric value

Temperature

Temperature during the week, in degrees Fahrenheit, a numeric value

Fuel_Price

Cost of fuel in the region, a numeric value

MarkDown1

Markdown 1, a numeric value

MarkDown2

Markdown 2, a numeric value

MarkDown3

Markdown 3, a numeric value

MarkDown4

Markdown 4, a numeric value

MarkDown5

Markdown 5, a numeric value

CPI

Consumer Price Index, a numeric value

Unemployment

Unemployment rate, a numeric value

Details

A dataset containing weekly sales data for a retail company, including information on holidays, markdowns, and economic indicators.

Examples

data(walmart_sales_weekly)
head(walmart_sales_weekly)
#> # A tibble: 6 × 17
#>   id    Store  Dept Date       Weekly_Sales IsHoliday Type    Size Temperature
#>   <fct> <dbl> <dbl> <date>            <dbl> <lgl>     <chr>  <dbl>       <dbl>
#> 1 1_1       1     1 2010-02-05       24924. FALSE     A     151315        42.3
#> 2 1_1       1     1 2010-02-12       46039. TRUE      A     151315        38.5
#> 3 1_1       1     1 2010-02-19       41596. FALSE     A     151315        39.9
#> 4 1_1       1     1 2010-02-26       19404. FALSE     A     151315        46.6
#> 5 1_1       1     1 2010-03-05       21828. FALSE     A     151315        46.5
#> 6 1_1       1     1 2010-03-12       21043. FALSE     A     151315        57.8
#> # ℹ 8 more variables: Fuel_Price <dbl>, MarkDown1 <dbl>, MarkDown2 <dbl>,
#> #   MarkDown3 <dbl>, MarkDown4 <dbl>, MarkDown5 <dbl>, CPI <dbl>,
#> #   Unemployment <dbl>